Put helper in its own file
