+++ Client/Frontend/Widgets/SiteTableViewController.swift
@@ -224,4 +224,4 @@ class SiteTableViewController: UIViewController, UITableViewDelegate, UITableVie
             tableView.reloadSections(IndexSet(rows.map { $0.section }), with: .none)
         }
     }
-//}
+}
